+++ fugue/src/test/java/com/atlassian/fugue/IterablesZipTest.java
@@ -27,18 +27,17 @@ import org.hamcrest.Matcher;
 import org.junit.Test;
 
 public class IterablesZipTest {
-  @SuppressWarnings("unchecked") @Test public void zipFromLists() {
-    assertThat(zip(asList(4, 3, 2, 1), asList(1, 2, 3, 4)), contains(pair(4, 1), pair(3, 2), pair(2, 3), pair(1, 4)));
+  @Test public void zipFromLists() {
+    assertThat(zip(asList(4, 3, 2, 1), asList(1, 2, 3, 4)), genericContains(pair(4, 1), pair(3, 2), pair(2, 3), pair(1, 4)));
   }
 
   @SuppressWarnings("unchecked") @Test public void zipFromLongerFirstList() {
-    assertThat(zip(asList(4, 2, 3, 1, 12), asList(1, 2, 3, 4)),
-      contains(pair(4, 1), pair(2, 2), pair(3, 3), pair(1, 4)));
+    assertThat(zip(asList(4, 2, 3, 1, 12), asList(1, 2, 3, 4)), genericContains(pair(4, 1), pair(2, 2), pair(3, 3), pair(1, 4)));
   }
 
   @SuppressWarnings("unchecked") @Test public void zipFromLongerLastList() {
     assertThat(zip(asList(4, 3, 2, 1), asList(6, 2, 5, 4, 5, 6)),
-      contains(pair(4, 6), pair(3, 2), pair(2, 5), pair(1, 4)));
+      genericContains(pair(4, 6), pair(3, 2), pair(2, 5), pair(1, 4)));
   }
 
   @Test public void zipWithFrom() {
@@ -47,8 +46,14 @@ public class IterablesZipTest {
 
   @Test public void testZipWithIndex() {
     @SuppressWarnings("unchecked")
-    Matcher<Iterable<? extends Pair<String, Integer>>> containsPairs = contains(pair("a", 0), pair("b", 1),
+    final Matcher<Iterable<? extends Pair<String, Integer>>> containsPairs = contains(pair("a", 0), pair("b", 1),
       pair("c", 2));
     assertThat(zipWithIndex(asList("a", "b", "c")), containsPairs);
+  }
+
+  @SuppressWarnings("unchecked")
+  @SafeVarargs
+  private final <A> Matcher<Iterable<? extends A>> genericContains(final A... as){
+    return contains(as);
   }
 }